+++ suite-common/suite-sync-evolu/src/createEvoluErrorHandler.ts
@@ -1,6 +1,7 @@
 import { Evolu } from '@evolu/common';
 
-import { SuiteSyncErrorHandler , asSuiteSyncOwnerId } from '@suite-common/suite-sync-types';
+import { asSuiteSyncOwnerId } from '@suite-common/suite-sync-storage';
+import { SuiteSyncErrorHandler } from '@suite-common/suite-sync-types';
 
 export const createEvoluErrorHandler =
     (evolu: Evolu<any>, errorHandler: SuiteSyncErrorHandler) => () => {
