+++ suite-common/suite-sync-evolu/src/createEvoluInstance.ts
@@ -2,7 +2,8 @@ import { Evolu, EvoluDeps, SimpleName, createEvolu } from '@evolu/common';
 import { sha256 } from '@noble/hashes/sha2.js';
 import { bytesToHex } from '@noble/hashes/utils.js';
 
-import { SuiteSyncErrorHandler , SuiteSyncOwner } from '@suite-common/suite-sync-types';
+import { SuiteSyncOwner } from '@suite-common/suite-sync-storage';
+import { SuiteSyncErrorHandler } from '@suite-common/suite-sync-types';
 
 import { createEvoluAppOwnerFromTrezorData } from './createEvoluAppOwnerFromTrezorData';
 import { createEvoluErrorHandler } from './createEvoluErrorHandler';
