+++ suite-common/suite-sync-evolu/src/evoluCreateSuiteSyncOwner.ts
@@ -1,5 +1,9 @@
-import { CreateSuiteSyncOwner, CreateSuiteSyncOwnerError } from '@suite-common/suite-sync-storage';
-import { asSuiteSyncOwnerId, asSuiteSyncOwnerSecretHex } from '@suite-common/suite-sync-types';
+import {
+    CreateSuiteSyncOwner,
+    CreateSuiteSyncOwnerError,
+    asSuiteSyncOwnerId,
+    asSuiteSyncOwnerSecretHex,
+} from '@suite-common/suite-sync-storage';
 import { err, ok } from '@trezor/type-utils';
 
 import { createEvoluAppOwnerFromTrezorData } from './createEvoluAppOwnerFromTrezorData';
