+++ suite-common/suite-sync-quota-manager/package.json
@@ -15,6 +15,7 @@
         "@reduxjs/toolkit": "2.11.2",
         "@suite-common/delegated-identity-key": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
+        "@suite-common/suite-sync-storage": "workspace:*",
         "@suite-common/suite-sync-types": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
         "@suite-common/suite-utils": "workspace:*",
