+++ suite-common/suite-sync-quota-manager/src/tests/ensureOwnerHasAllocatedQuotaThunk.test.ts
@@ -1,6 +1,6 @@
 import { mocked } from 'jest-mock';
 
-import { asSuiteSyncOwnerId } from '@suite-common/suite-sync-types';
+import { asSuiteSyncOwnerId } from '@suite-common/suite-sync-storage';
 import { asDelegatedIdentityKey } from '@suite-common/suite-types';
 import { WalletDescriptor, asWalletDescriptor } from '@suite-common/wallet-types';
 import { err, ok } from '@trezor/type-utils';
