+++ suite-common/suite-sync-storage/tsconfig.json
@@ -5,7 +5,6 @@
         "exactOptionalPropertyTypes": true
     },
     "references": [
-        { "path": "../suite-sync-types" },
         { "path": "../wallet-config" },
         { "path": "../wallet-types" },
         { "path": "../../packages/type-utils" }
