+++ suite-common/suite-sync-types/src/SuiteSyncErrorHandler.ts
@@ -1,6 +1,6 @@
 import { Dispatch } from '@reduxjs/toolkit';
 
-import { SuiteSyncOwnerId } from './owner/suiteSyncOwner';
+import { SuiteSyncOwnerId } from '@suite-common/suite-sync-storage';
 
 export type CreateSuiteSyncErrorHandlerDep = {
     dispatch: Dispatch;
