+++ suite-common/suite-sync-types/src/quotaManager/ensureOwnerHasAllocatedQuotaThunk.ts
@@ -1,8 +1,8 @@
+import { SuiteSyncOwnerId } from '@suite-common/suite-sync-storage';
 import { DelegatedIdentityKey } from '@suite-common/suite-types';
 import { WalletDescriptor } from '@suite-common/wallet-types';
 import { Result } from '@trezor/type-utils';
 
-import { SuiteSyncOwnerId } from '../owner/suiteSyncOwner';
 import { WriteModeRequiredForAllocationErrType } from '../refreshSuiteSyncKeys';
 
 export type HttpErrType = { type: 'HttpError' };
