+++ suite-common/suite-sync/src/owner/createRetrieveSuiteSyncOwner.ts
@@ -1,7 +1,10 @@
 import { getProofOfDelegatedIdentity } from '@suite-common/delegated-identity-key';
 import { ProofOfDelegatedSignFailedType } from '@suite-common/delegated-identity-key-types';
-import { CreateSuiteSyncOwner, CreateSuiteSyncOwnerError } from '@suite-common/suite-sync-storage';
-import { SuiteSyncOwner } from '@suite-common/suite-sync-types';
+import {
+    CreateSuiteSyncOwner,
+    CreateSuiteSyncOwnerError,
+    SuiteSyncOwner,
+} from '@suite-common/suite-sync-storage';
 import { DelegatedIdentityKey, TrezorDeviceWithState } from '@suite-common/suite-types';
 import { DeviceError } from '@suite-common/wallet-core';
 import { DeviceErrorType } from '@suite-common/wallet-types';
