+++ suite-common/suite-sync/src/tests/createRefreshSuiteSync.test.ts
@@ -2,11 +2,8 @@ import { Dispatch } from '@reduxjs/toolkit';
 
 import { EnsureDelegatedIdentityKey } from '@suite-common/delegated-identity-key-types';
 import { mockNotExpected } from '@suite-common/dependency-injection';
-import {
-    EnsureSuiteSyncOwner,
-    asSuiteSyncOwnerId,
-    asSuiteSyncOwnerSecretHex,
-} from '@suite-common/suite-sync-types';
+import { asSuiteSyncOwnerId, asSuiteSyncOwnerSecretHex } from '@suite-common/suite-sync-storage';
+import { EnsureSuiteSyncOwner } from '@suite-common/suite-sync-types';
 import { TrezorDevice, asDelegatedIdentityKey } from '@suite-common/suite-types';
 import { ok } from '@trezor/type-utils';
