+++ suite/e2e/tsconfig.json
@@ -19,7 +19,7 @@
             "path": "../../suite-common/suite-sync-evolu"
         },
         {
-            "path": "../../suite-common/suite-sync-types"
+            "path": "../../suite-common/suite-sync-storage"
         },
         {
             "path": "../../suite-common/suite-types"
